fix: import usage check could pass on a less-specific import
For example, the datatype `foo.bar.Baz` would pass the import
check for either `foo` or `foo.bar`.

This becomes an issue because the `[]string packageNames` is
non-determinstically ordered and can cause flaky checks. When the
less-specific import is used for a more-specific datatype, the
more-specific import "becomes" unused later in the check.

This change means that an import is only considered used if
one or more types use the fully qualified import (everything up
to the most-specific '.' separator).

In order to support nested-message imports like:

    foo.bar.BazMessage.InnerMessage

the "most specific" check uses the last package segment that is
not uppercased as the last package segment. This aligns with
naming conventions laid out by Google and most common usage.