fixes #12: Cannot find messages when package is defined across more than one dependency
This changes protoFileOracle to store a slice of pointers to ProtoFiles
rather than a single reference.

When a package was defined across multiple dependencies,
`parseDependencies` would only store a reference to the first protoFile
found with the same package name. This resulted in verification errors
for any messages and enums defined in subsequent imports with the same
package name.